added date utility class & methods
