+++ app/src/main/res/layout/activity_food_reminder.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="edu.bu.fitnessfriend.fitnessfriend.food_reminder"
+    tools:context="edu.bu.fitnessfriend.fitnessfriend.food.food_reminder"
     tools:layout_editor_absoluteY="81dp"
     tools:layout_editor_absoluteX="0dp">
